fix: improve login view appearance
* Use consistent text sizing.
* Use Typescript
* Use react-hook-form.
